refactor: restructure gui module and extract telescope picker to its own file
This is to make it easier to add new pickers
Additionally, the entrypoint of search has changed from `search.New` to
`gui.open`, because `run_search` is now passed a picker object
