+++ lua/venv-selector/user_commands.lua
@@ -1,5 +1,4 @@
 local config = require("venv-selector.config")
-local search = require("venv-selector.search")
 local gui = require("venv-selector.gui")
 local log = require("venv-selector.logger")
 
@@ -7,8 +6,7 @@ local M = {}
 
 function M.register()
     vim.api.nvim_create_user_command("VenvSelect", function(opts)
-        gui.open(search.search_in_progress)
-        search.New(opts)
+        gui.open(opts)
     end, { nargs = "*", desc = "Activate venv" })
 
     vim.api.nvim_create_user_command("VenvSelectLog", function()
